Retain the default-directory in eless
If you did "grep --help | eless"

the "grep --help" output was getting saved to /tmp dir and then that was opened
in emacs, so that changed the default-directory to "/tmp/".

That caused annoyance when doing "C-x C-f" and such as it would set the starting
directory to "/tmp/".

This commit fixes that.

Now the default-directory is set to the directory from which eless is launched.