Fix over-eager view-mode buffer reverting
In the case of multiple buffers opened in view-mode, only the current buffer, if
in view-mode, will be force-reverted. This applies only when calling
eless/kill-emacs-or-buffer command.

When C-x C-c is pressed, all view-mode buffers will still be force-reverted. Now
C-x C-c is bound to the new eless/save-buffer-maybe-and-kill-emacs command added
in this commit.

This further improves the behavior explained in https://github.com/kaushalmodi/eless/issues/7