Support opening archives
    eless foo.tar.xz

With this commit, now instead of always enabling view-mode and then disabling it
when one of the "special" major modes like dired, Info, Man and tar is detected,
we instead wait to check if the major mode is none of the above and only then
enable view-mode.

This optimizes things a bit; now 'view feature is not loaded unnecessarily.

For example, 'view is not loaded if doing "eless foo.tar.xz" or "eless .".